moved compare_state_dict to tensor_utils
