+++ muutils/tensor_utils.py
@@ -336,3 +336,15 @@ def rpad_array(
 ) -> np.ndarray:
     """pad a 1-d array on the right with pad_value to length `pad_length`"""
     return pad_array(array, pad_length, pad_value, rpad=True)
+
+
+def compare_state_dicts(d1: dict, d2: dict):
+    assert d1.keys() == d2.keys(), "state dict keys don't match!"
+    keys_failed: list[str] = list()
+    for k, v in d1.items():
+        v_load = d2[k]
+        if not (v == v_load).all():
+            keys_failed.append(k)
+    assert (
+        len(keys_failed) == 0
+    ), f"{len(keys_failed)} / {len(d1)} state dict elements don't match: {keys_failed}"
